pyproject.toml
[project]
name = "agency"
version = "0.1.0"
description = "Automated multi-step workflows for AI coding agents."
requires-python = ">=3.14"
dependencies = ["PyYAML"]

[project.scripts]
agency = "agency:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "import-linter>=2.11",
    "inline-snapshot[dirty-equals]>=0.32.5",
    "prek>=0.3.8",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "ty>=0.0.25",
]
